+++ tigergraph_mcp/tools/db/data_source/__init__.py
@@ -6,14 +6,20 @@
 # under the License. The software is provided "AS IS", without warranty.
 
 from .create_data_source_tool import create_data_source
+from .update_data_source_tool import update_data_source
 from .get_data_source_tool import get_data_source
 from .drop_data_source_tool import drop_data_source
+from .get_all_data_sources_tool import get_all_data_sources
+from .drop_all_data_sources_tool import drop_all_data_sources
 from .preview_sample_data_tool import preview_sample_data
 
 
 __all__ = [
     "create_data_source",
+    "update_data_source",
     "get_data_source",
     "drop_data_source",
+    "get_all_data_sources",
+    "drop_all_data_sources",
     "preview_sample_data",
 ]